[TEMPORARY] Fix project dependencies to allow Behat to run with PHP 7.3
This was the product of running `composer update` on the project, as it seems the current composer.lock has issues with the current PHP version (7.3).

Ideally we could identify the specific issue and fix only it, but again, this PR is for illustrative purposes so this is just a practical approach to show that the changes in this PR would work if we decide to move forward with it.