Fix composer dependencies issues
The current composer.lock causes an error "Your lock file does not contain a compatible set of packages" when installing composer dependencies.

Since this is a package and not an application, composer.lock doesn't make sense. Ideally we would test the package with the lowest and highest version of all supported dependencies so we confirm it fully supports the range in composer.json, but doing that would be out of the scope of this PR.

I am not even sure CI is currently fully working for this package, but all tests pass locally.