Add automated testing to environment variable override.
This adds a test scenario to confirm the new functionality indeed overrides the configuration value.

If we expose the BEHAT_RETRY_TIMEOUT environment variable on the PHP container it is then available to the Behat child process run on the test.

For this reason, I added a step specifically for setting this environment variable, as that is the only one exposed to the Behat process currently by being proxied by the PHP container to the behat child process, and have added a BeforeScenario hook for clearing only this specific environment variable on the behat parent process to isolate tests.