Support overriding timeout through by environment variable. (#33)
* Add environment variable support for retry timeout configuration

This is not tested yet due to requiring changes to a 3rd-party dependency to allow us to test it (or using a different package for testing). This PR aims to enable the functionality and handle coordinating with the dependency maintainers on a follow-up issue (https://github.com/Medology/Umami/issues/2862)

* Clarify usage of BEHAT_RETRY_TIMEOUT environment variable in README

* Fix composer dependencies issues

The current composer.lock causes an error "Your lock file does not contain a compatible set of packages" when installing composer dependencies.

Since this is a package and not an application, composer.lock doesn't make sense. Ideally we would test the package with the lowest and highest version of all supported dependencies so we confirm it fully supports the range in composer.json, but doing that would be out of the scope of this PR.

I am not even sure CI is currently fully working for this package, but all tests pass locally.

* Add automated testing to environment variable override.

This adds a test scenario to confirm the new functionality indeed overrides the configuration value.

If we expose the BEHAT_RETRY_TIMEOUT environment variable on the PHP container it is then available to the Behat child process run on the test.

For this reason, I added a step specifically for setting this environment variable, as that is the only one exposed to the Behat process currently by being proxied by the PHP container to the behat child process, and have added a BeforeScenario hook for clearing only this specific environment variable on the behat parent process to isolate tests.

* CodeRabbit CR: improve validation and add testing

* Add extra scenario to validate environment variable overrides both if higher and lower than timeout

* Use bin/php to run tests

To share the environment variable with the Behat run in the child process, we need to use bin/php

* Update CircleCI configuration to use chekote/php image to match local environment

* Install composer on CircleCI job

* Use machine executor on CircleCI

* Use bin/php to invoke composer